move NotificationForConvoOption declaration + getConversationNotificationSettingType created as used twice
