+++ ts/components/session/menu/ConversationHeaderMenu.tsx
@@ -13,11 +13,13 @@ import {
   getLeaveGroupMenuItem,
   getMarkAllReadMenuItem,
   getNotificationForConvoMenuItem,
+  getPinConversationMenuItem,
   getRemoveModeratorsMenuItem,
   getUpdateGroupNameMenuItem,
 } from './Menu';
-import { NotificationForConvoOption, TimerOption } from '../../conversation/ConversationHeader';
+import { TimerOption } from '../../conversation/ConversationHeader';
 import { ConversationNotificationSettingType } from '../../../models/conversation';
+import { NotificationForConvoOption } from '../../../state/ducks/conversations';
 
 export type PropsConversationHeaderMenu = {
   conversationId: string;
